Added npm react-native-cli command to windows and linux
Summary:
The windows and linux block didn't have the command to install react-native-cli.

The command is the exact same as it is on macOS. It makes the guide a little confusing because it leaves out this critical step and it's difficult to understand whether or not this needed component was already indirectly installed in one of the other steps.
Closes https://github.com/facebook/react-native/pull/11135

Differential Revision: D4236984

Pulled By: mkonicek

fbshipit-source-id: dca237132df7d80b5a02eaf09faec53e13e005cb