IntegrationWithExsistingApps Android Update
Summary:
Added permission code to handle the user permission accepted or denied cases which causes potential app crashing if not handled. Only a small piece of code added to the guide to resolve an issue I was facing with permission.

The app that I was testing following the guide available, it was crashing showing permission denied error on the android monitor, and there was no step for how to resolve the issue. So I tried out a solution so thought would share it on the official Guide to help people who are stuck with the same issue.
Closes https://github.com/facebook/react-native/pull/11687

Differential Revision: D4377882

Pulled By: ericvicenti

fbshipit-source-id: a7db104c7c1b7e1fb2590c83118715840c17927b