Make a section on keyboard shortcuts more prominent
Summary:
The note is currently in "Reloading JavaScript". However I was struggling with the first step ("Accessing the In-App Developer Menu") so didn't read down to that section.

I think it makes sense to make this the very first thing the reader sees when visiting the page.
Closes https://github.com/facebook/react-native/pull/12828

Differential Revision: D4681842

Pulled By: hramos

fbshipit-source-id: 1afbec7f62a2ec97febdd05987e7fc1541d3ee28